added mapper

+++ src/main/java/com/senin/admissions_committee_servlet/DAO/mapper/ApplicantProfileMapper.java
@@ -25,7 +25,7 @@ public class ApplicantProfileMapper implements ObjectMapper<ApplicantProfile> {
     }
 
     @Override
-    public ApplicantProfile makeUnique(Map<Integer, ApplicantProfile> cache, ApplicantProfile teacher) {
+    public ApplicantProfile makeUnique(Map<Long, ApplicantProfile> cache, ApplicantProfile teacher) {
         return null;
     }
 }
