+++ src/main/java/com/senin/admissions_committee_servlet/DAO/mapper/ObjectMapper.java
@@ -8,6 +8,5 @@ public interface ObjectMapper<T> {
 
     T extractFromResultSet(ResultSet rs) throws SQLException;
 
-    T makeUnique(Map<Integer, T> cache,
-                 T teacher);
+    T makeUnique(Map<Long, T> cache,T teacher);
 }
